+++ cxl/acpi/rdpas.c
@@ -1,0 +1,14 @@
+// SPDX-License-Identifier: GPL-3.0
+// Copyright (C) 2026 Rong Tao
+#include <stdio.h>
+#include <stdlib.h>
+#include <malloc.h>
+#include <math.h>
+#include "rdpas.h"
+
+void display_rdpas(struct rdpas *rdpas)
+{
+	printf("---------------- RDPAS ----------------\n");
+	/* TODO: */
+	printf("TODO...\n");
+}