+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to
 #### Breaking Changes
 - add .gitconfig.sh in rootdir;
 #### Added
-- cxl/acpi: add cxims.c;
+- cxl/acpi: add cxims.c, rdpas.c;
 - acpi: add table subdir;
 - bpf: test kfunc `bpf_rcu_read_lock()`/`bpf_rcu_read_unlock()`;
 - macro: add `__powerpc64__.c`;